// src/app/model/utilisateur.ts
export class Utilisateur
{
     ID_UTILISATEUR: number;
     ID_TYPEUTILISATEUR: number;
     
     MAIL_UTILISATEUR: string;
     NOM_UTILISATEUR: string;
     PRENOM_UTILISATEUR: string;
     TEL_UTILISATEUR: string;
     PSEUDO_UTILISATEUR: string;
     MOTDEPASSE_UTILISATEUR: string;
     NIVEAU_UTILISATEUR: number;
     EXPERIENCE_UTILISATEUR: number;

     IMAGE_UTILISATEUR: string;

    constructor() {}
}